Fix late handler registering
We did not execute late registered events if viewport is
not at default level (as promised in the documentation).
Ensure to call new onChange handlers right away when late
registered. Of course only if the viewport is not at the
default level. Also make sure to not invoke the handler
twice when initial resize event is pending.